sec={Article.Sec}  more concise.
I also made a change to the Z/paras widget of Cmacc-Paris2.  I changed
blist=…. (in Z/ol/Base) to add a <br> at the end.
For some reason that I don’t understand, a <ul> list is spaced at the
top but not at the bottom.  (An <ol> is not spaced at either top or
bottom.)  Perhaps this can be fixed in CSS.  I won’t roll this out to
CmaccDoc_Z (the general source for Z) until we get some experience.